Fix an issue where updated player pic was not displayed after download
When an updated player profile picture was downloaded, the new picture
would be displayed in some places, and the old picture would be
displayed in other places. This was happening because the smaller sizes
of the old picture were not being removed from the image cache after the
newer picture was downloaded.